scripts/requirements.txt
tailk
homebrew-pypi-poet==0.10.0
